Develop v1 to main

+++ TaskPlannerMetrum/Model/OrderManagementInfo.cs
@@ -9,9 +9,9 @@ namespace TaskPlannerMetrum.Model
 
         public int ContractID { get; set; }
 
-        public decimal? TotalCostHoursPlanned { get; set; }
+        public double? TotalCostHoursPlanned { get; set; }
 
-        public decimal? TotalCostHoursExecuted { get; set; }
+        public double? TotalCostHoursExecuted { get; set; }
 
         public double? TotalDifferenceExpectedExecuted { get; set; }
